Fix custom Dbm scripts
The custom database migration scripts in our `scripts` directory had
been broken, apparently since somewhere around the FOSS release.  It is
still unclear how this happened, but this commit fixes it by changing
the way in which the database-migration plugin's base directory is
found.

NOTE: The new method does require the plugin to actually be installed.
It will not work in a fresh clone of the code, but will work after
simply running `grails compile`